Increased timeout for item commands and updates to 30 seconds (#4679)
* Increased timeout for item commands and updates to 30 seconds

+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/internal/common/SafeCallerBuilderImpl.java
@@ -33,7 +33,7 @@ public class SafeCallerBuilderImpl<T> implements SafeCallerBuilder<T> {
 
     private final T target;
     private final Class<?>[] interfaceTypes;
-    private int timeout;
+    private long timeout;
     private Object identifier;
     @Nullable
     private Consumer<Throwable> exceptionHandler;
@@ -66,7 +66,7 @@ public class SafeCallerBuilderImpl<T> implements SafeCallerBuilder<T> {
     }
 
     @Override
-    public SafeCallerBuilder<T> withTimeout(int timeout) {
+    public SafeCallerBuilder<T> withTimeout(long timeout) {
         this.timeout = timeout;
         return this;
     }
